Add files via upload
Sentiment Analysis: Sentiment analysis is the process of determining the emotional tone of a piece of text, such as a review or a social media post. In this project, we will use NLP techniques to analyze the sentiment. Here in one project we have analyzed the sentiment by giving a text by the user and the other program includes sentimental analysis of an article from the internet.